total=836,084 (final)
Well, probably final actually. The lead maintainer of bruteforcemovable.com has retired so the influx of msed data has ceased.

This is okay since 1) we have a bunch of data already, and 2) the manufacturing of 3ds has stopped months ago, so we have an accurate snapshot of 3ds movables for the long run. The current and final average offset to msed hits are: old3ds=1.7 and new3ds=1.4. This is a little bit less than a 2^34 keyspace bruteforce.

Many thanks to Deadphoenix and the numerous bruteforce helpers for all their work in maintaining bfm over the years (and will continue to do so with the alt site: https://seedminer.hacks.guide/).